- changed json library; - archive file now also contains default serialization objs; - minor adjustments to factory objects - added .gfl and test factory

+++ Engine/Gumball/Source/Gumball/Timer.cpp
@@ -7,7 +7,7 @@ Timer::Timer() {
     time = high_resolution_clock::now();
 }
 void Timer::start() {
-    milliseconds = duration_cast<std::chrono::milliseconds>(high_resolution_clock::now() - time).count();
-    seconds = milliseconds / 1000.0;
+    milliseconds = (double)duration_cast<std::chrono::milliseconds>(high_resolution_clock::now() - time).count();
+    seconds = (float)milliseconds / 1000.0;
     time = high_resolution_clock::now();
 }
